requirements.txt
retina-face>=0.0.13
numpy>=1.23.5
tensorflow
keras>=2.10.0
opencv-python>=4.6.0
timm>=0.6.13
torch>=2.0.0
torchvision
pykalman>=0.9.5
scipy>=1.10.0
tslearn>=0.5.3.2
protobuf>=4.22.1
requests>=2.28.1
gdown
grad-cam
matplotlib
moviepy
pydub
setuptools==65.5.1
wrapt
huggingface_hub
onnx>=1.7.0
python-dateutil
ffmpeg-python
wget>=3.2
sox>=1.4
ffmpeg>=1.4
text-unidecode>=1.3
numba>=0.57.0
nemo_toolkit
Cython
noisereduce
pandas
sacremoses>=0.0.43
sentencepiece<1.0.0
youtokentome>=1.0.5
inflect
braceexpand
webdataset
editdistance
g2p_en
inflect
ipywidgets
jiwer
kaldi-python-io
kaldiio
librosa>=0.9.0
marshmallow
matplotlib
packaging
pyannote.core
pyannote.metrics
ruamel.yaml
soundfile
texterrors
hydra-core>=1.2.0,<1.3
omegaconf>=2.2,<2.3
pytorch-lightning>=1.9.0,<=1.9.4
pyyaml<6  # Pinned until omegaconf works with pyyaml>=6
torchmetrics>=0.11.0
transformers>=4.0.1
wandb
webdataset>=0.1.48,<=0.1.62